+++ index.d.ts
@@ -1,4 +1,3 @@
-export * from './lnd_gateway';
 export * from './lnd_grpc';
 export * from './lnd_methods';
 export * from './typescript';