+++ lnd_gateway/index.d.ts
@@ -1,3 +1,0 @@
-export * from './emit_grpc_events';
-export * from './grpc_router';
-export * from './lnd_gateway';