+++ lnd_gateway/index.js
@@ -1,5 +1,0 @@
-const emitGrpcEvents = require('./emit_grpc_events');
-const grpcRouter = require('./grpc_router');
-const lndGateway = require('./lnd_gateway');
-
-module.exports = {emitGrpcEvents, grpcRouter, lndGateway};