+++ test/typescript/grpc_router.test-d.ts
@@ -1,8 +1,0 @@
-import {expectType} from 'tsd';
-import {Router} from 'express';
-import {grpcRouter} from '../../lnd_gateway';
-
-expectType<Router>(grpcRouter({}));
-expectType<Router>(grpcRouter({cert: '00'}));
-expectType<Router>(grpcRouter({socket: 'socket'}));
-expectType<Router>(grpcRouter({cert: '00', socket: 'socket'}));